Changed request timeout value for H5P requests
Made sure to override timeout value for H5P cURL requests. The values is
set to 90 if the request is made to api.h5p.org, has timeout and timeout
value is less than 90 seconds. 90 seconds seems to be enough to fetch
the libraries in situations with limited or slow connection speed.