Added a standalone class for admin related functionality.
Moved some logic to the newly created admin class. Some uncommented
code would be usef for still not fully implemented logic. Http request
timeout extension was moved to the admin part as the libraries importing
requests should be made within an admin context. Code could still be
unstable and should not be used in production.