Fixes batches and statement AJAX logic.
Batches code had a few issues which resulted in multiple duplicated
being created. Statement AJAX did not run JSON.stringify() on the
payload, which resulted in numbers and booleans being converted into
trings. Introduced the change to the codebase to prevent that from
happening in the future. Implemented a method that would check the
statement and fix these issues. Fix is being applied to each statement
befoe adding it to the batch. NB! Code needs checking and verson bump
will be made once the stability is ensured. Constantly applying fix
might not be the best solution and a migration to existing statements
data could be implemented instead.